update girder and osumo
This update opens the doors for public users to create accounts and log
in using those account, and to run analysis with their data.  Currently,
however, one would still need to browse the girder UI to manage their
data sets.

This should be the final deployment update before the architectural
rework.  It will also probably be the last deployment stuck using Girder
1.0.

 - girder -> latest 1.x-maintenance branch
 - osumo -> latest master